% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enter.order.R
\name{enter.order}
\alias{enter.order}
\title{Entering order of predictors in a regularized path}
\usage{
enter.order(beta, eps = 1e-06)
}
\arguments{
\item{beta}{a matrix with each column being the norm of coefficients
for a given tuning parameter \code{lambda}. The number
of columns is \code{length(lambda)}.}

\item{eps}{tolerance when extracting non-zero coefficients. Default value is
\code{1e-6}.}
}
\value{
A vector containing the indexes of predictors following their entering order.
}
\description{
This function allows you to obtain the entering order of predictors by a given regularized
solution path.
}
\examples{
x = matrix(rnorm(100 * 10), 100, 10)
y = rnorm(100)
library(glmnet)
fit1 = glmnet(x, y)
enter.order(fit1$beta)
}
