% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Generate_Y.R
\name{Generate_Y}
\alias{Generate_Y}
\title{Simulate response variable from a glm family}
\usage{
Generate_Y(predy, sigmasq = 1, n, family = "gaussian")
}
\arguments{
\item{predy}{Values of a linear predictor that can be written as \eqn{X\beta}.}

\item{sigmasq}{Variance of errors for a Gaussian linear model. Default value is 1.}

\item{n}{sample size}

\item{family}{response type. A character string representing one of
the families: \code{"gaussian"}, \code{"binomial"}, \code{"gam"}, \code{"poission"}.
Default is \code{"gaussian"}.}
}
\value{
Returns a vector of length \code{n} with elements drawn from
a specified family
}
\description{
This is a generic function that can simulate new observations from a
distribution in exponential family.
}
\examples{
set.seed(0)
n=50
p=10
X = matrix(rnorm(n*p),nrow=n,ncol=p)
true_b = c(1:3, rep(0,p-3))
predy = X \%*\% true_b
#Simulate obs from Gaussian linear model
Generate_Y(predy, n)
}
\keyword{glm}
\keyword{simulation}
