% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_NMCS.R
\name{output_NMCS}
\alias{output_NMCS}
\title{Summary of nested model confidence sets}
\usage{
output_NMCS(nmcs.r, alpha, predictors = NULL)
}
\arguments{
\item{nmcs.r}{a result of \code{NMCS}.}

\item{alpha}{A sequence of significance levels. Default value is 0.05.}

\item{predictors}{The indexes of all predictors. Default value is \code{1:p}.}
}
\value{
Returns a list including the predictors indexes of selected model, and a
dataframe including all model confidence sets for user-given \code{alpha} levels.
\item{hat_M}{indexes of predictors in the selected model.}
\item{MCS.frame}{a dataframe containing the information about model confidence sets.
\describe{
\item{CL}{confidence levels.}
\item{bcp}{the bootstrap coverage probabilities.}
\item{width}{width of NMCS, which equals the size difference between lbm and ubm.}
\item{lbm}{lower bound models.}
\item{ubm}{upper bound models.}
}
}
}
\description{
This is a generic function used to produce result summaries of nested model confidence sets.
}
\examples{
set.seed(0)
n=50
p=10
X = matrix(rnorm(n*p),nrow=n,ncol=p)
true_b = c(1:3, rep(0,p-3))
Y = X \%*\% true_b + rnorm(n)
alpha=c(0.05,0.1)
result=NMCS(Y, X, alpha=alpha)
output_NMCS(result,alpha=alpha)
}
